+++ drizzle.config.ts
@@ -1,10 +1,12 @@
+import { config } from 'dotenv';
+config({ path: '.env.local' });
 import type { Config } from 'drizzle-kit';
 
 export default {
   schema: './db/schema.ts',
   out: './drizzle',
-  dialect: 'sqlite',
+  dialect: 'postgresql',
   dbCredentials: {
-    url: 'budget.db',
+    url: process.env.DATABASE_URL!,
   },
 } satisfies Config;